--- frontend/src/shared/index.js
@@ -1,4 +1,0 @@
-
-import SignInPage from './components/SignInPage'
-
-export {SignInPage}
